printf

+++ ft_hex.c
@@ -3,25 +3,25 @@
 /*                                                        :::      ::::::::   */
 /*   ft_hex.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: cmichez <cmichez@student.42.fr>            +#+  +:+       +#+        */
+/*   By: cmichez <cmichez@student.42nice.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/12/02 13:45:25 by cmichez           #+#    #+#             */
-/*   Updated: 2022/12/02 13:45:25 by cmichez          ###   ########.fr       */
+/*   Created: 2022/12/12 16:05:43 by cmichez           #+#    #+#             */
+/*   Updated: 2022/12/12 16:05:43 by cmichez          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int	ft_put_hex(long n, char *base)
+int	ft_put_hex(int n, char *base)
 {
-	int i;
+	int	i;
 
 	i = 0;
-	ft_conv_hex(n, base, &i);
+	ft_conv_hex((unsigned int) n, base, &i);
 	return (i);
 }
 
-void ft_conv_hex(long n, char *base, int *i)
+void	ft_conv_hex(unsigned int n, char *base, int *i)
 {
 	if (n >= 16)
 		ft_conv_hex(n / 16, base, i);
@@ -29,7 +29,7 @@ void ft_conv_hex(long n, char *base, int *i)
 	(*i)++;
 }
 
-void ft_ptr_hex(unsigned long nb, char *base, int *len)
+void	ft_ptr_hex(unsigned long nb, char *base, int *len)
 {
 	if (nb >= 16)
 		ft_ptr_hex(nb / 16, base, len);
@@ -37,9 +37,9 @@ void ft_ptr_hex(unsigned long nb, char *base, int *len)
 	(*len)++;
 }
 
-int ft_print_ptr(unsigned long ptr)
+int	ft_print_ptr(unsigned long ptr)
 {
-	int len;
+	int	len;
 
 	len = 0;
 	len += ft_putstr("0x");
